Refactoring, part II
