+++ src/ua/spalah/bank/enums/Sex.java
@@ -1,4 +1,4 @@
-package ua.spalah.bank;
+package ua.spalah.bank.enums;
 
 /**
  * Created by root on 03.01.2017.
