+++ src/ua/spalah/bank/model/CheckingAccount.java
@@ -1,6 +1,6 @@
 package ua.spalah.bank.model;
 
-import ua.spalah.bank.AccountType;
+import ua.spalah.bank.enums.AccountType;
 
 /**
  * Created by root on 23.12.2016.
