+++ src/ua/spalah/bank/model/Client.java
@@ -1,7 +1,7 @@
 package ua.spalah.bank.model;
 
 import ua.spalah.bank.Account;
-import ua.spalah.bank.Sex;
+import ua.spalah.bank.enums.Sex;
 import java.util.ArrayList;
 import java.util.List;
